# lib/api-zod/src/generated/types/alert.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * BH Farm OS operations API
 * OpenAPI spec version: 0.1.0
 */

export interface Alert {
  id: number;
  severity: string;
  title: string;
  detail: string;
  category: string;
  /** @nullable */
  dueLabel: string | null;
}