# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * BH Farm OS operations API
 * OpenAPI spec version: 0.1.0
 */
import type { Activity } from './activity';
import type { Alert } from './alert';
import type { LivestockSummary } from './livestockSummary';
import type { Task } from './task';
import type { WaterSummary } from './waterSummary';

export interface DashboardSummary {
  farmName: string;
  location: string;
  dateLabel: string;
  livestock: LivestockSummary;
  activeCropCycles: number;
  orchardTrees: number;
  water: WaterSummary;
  lowStockItems: number;
  tasksDueToday: number;
  overdueTasks: number;
  monthlyRevenue: number;
  monthlyExpenses: number;
  monthlyProfit: number;
  tasks: Task[];
  alerts: Alert[];
  recentActivity: Activity[];
}