# lib/api-zod/src/generated/types/enterprisePerformance.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * BH Farm OS operations API
 * OpenAPI spec version: 0.1.0
 */

export interface EnterprisePerformance {
  name: string;
  revenue: number;
  expenses: number;
  profit: number;
  margin: number;
}