# lib/api-zod/src/generated/types/farmZone.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * BH Farm OS operations API
 * OpenAPI spec version: 0.1.0
 */

export interface FarmZone {
  id: number;
  name: string;
  type: string;
  /** @nullable */
  dimensions: string | null;
  status: string;
  accent: string;
}