# lib/api-zod/src/generated/types/financeSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * BH Farm OS operations API
 * OpenAPI spec version: 0.1.0
 */
import type { EnterprisePerformance } from './enterprisePerformance';
import type { MonthlyFinance } from './monthlyFinance';

export interface FinanceSummary {
  revenue: number;
  expenses: number;
  profit: number;
  revenueChange: number;
  expenseChange: number;
  enterprisePerformance: EnterprisePerformance[];
  monthly: MonthlyFinance[];
}