# lib/api-zod/src/generated/types/goat.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * BH Farm OS operations API
 * OpenAPI spec version: 0.1.0
 */

export interface Goat {
  id: number;
  goatId: string;
  name: string;
  sex: string;
  breed: string;
  /** @nullable */
  dateOfBirth: string | null;
  /** @nullable */
  currentWeightKg: number | null;
  paddock: string;
  status: string;
  pregnancyStatus: string;
  /** @nullable */
  expectedKiddingDate: string | null;
  healthStatus: string;
  /** @nullable */
  daysRemaining: number | null;
}